[refactor] 데이터베이스 스키마 및 사용자 관련 임포트 경로 수정
