+++ src/lib/db/schema.ts
@@ -1,0 +1,12 @@
+import { pgTable, text, timestamp } from "drizzle-orm/pg-core";
+
+export const users = pgTable("user", {
+  id: text("id")
+    .primaryKey()
+    .$defaultFn(() => crypto.randomUUID()),
+  name: text("name"),
+  email: text("email").unique().notNull(),
+  emailVerified: timestamp("emailVerified", { mode: "date" }),
+  image: text("image"),
+  hashedPassword: text("hashedPassword"),
+});
